Split out modify collateral tests into deposit, withdraw and failure
